pyproject.toml
[project]
name = "monitoring-api"
dynamic = ["version"]
description = "Django API and data import/export Python software package for long-term environmental monitoring data."
authors = [
    {name = "Rebecca Kurup Buchholz", email = "rebecca.kurup@wsl.ch"},
]
dependencies = [
    "asgiref==3.4.1",
    "atomicwrites==1.4.0",
    "attrs==20.3.0",
    "certifi==2022.6.15",
    "cffi==1.14.4",
    "chardet==3.0.4",
    "colorama==0.4.4",
    "configparser==5.0.2",
    "cryptography==3.4.5",
    "datetime==4.3",
    "deepdiff==5.8.1",
    "django==3.2.14",
    "django-environ==0.9.0",
    "django-postgres-copy==2.5.0",
    "djangorestframework==3.12.2",
    "et-xmlfile==1.0.1",
    "idna==2.10",
    "iniconfig==1.1.1",
    "jdcal==1.4.1",
    "numpy==1.22.3",
    "ordered-set==4.1.0",
    "pandas==1.2.4",
    "pluggy==0.13.1",
    "postgres==3.0.0",
    "psycopg2-binary==2.8.6",
    "psycopg2-pool==1.1",
    "py==1.10.0",
    "pycparser==2.20",
    "pyopenssl==20.0.1",
    "pyparsing==2.4.7",
    "pytest==6.1.2",
    "python-dateutil==2.8.1",
    "pytz==2020.1",
    "requests==2.25.1",
    "six==1.15.0",
    "sqlparse==0.3.1",
    "toml==0.10.2",
    "tzdata==2021.5",
    "urllib3==1.26.5",
    "zope.interface==5.1.0",
    "gunicorn>=20.1.0",
]
requires-python = ">=3.8"
license = {text = "MIT"}
[project.optional-dependencies]

[project.urls]
homepage = "https://envidat.ch"
documentation = "https://gitlabext.wsl.ch/envidat/monitoring-backend"
repository = "https://gitlabext.wsl.ch/envidat/monitoring-backend"

[tool.pdm]
version = {from = "monitoring-api/project/__version__.py"}

[[tool.pdm.source]]
name = "pypi"
url = "https://pypi.org/simple"
verify_ssl = true


[tool.pdm.dev-dependencies]
dev = [
    "debugpy>=1.6.2",
]
[build-system]
requires = ["pdm-pep517"]
build-backend = "pdm.pep517.api"

[tool.commitizen]
name = "cz_conventional_commits"
version = "0.2.0"
version_files = [
    "pyproject.toml:version",
    "monitoring-api/project/__version__.py",
    ".env:APP_VERSION",
    "chart/Chart.yaml:appVersion",
]
